backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
python-dotenv==1.0.0
django-summernote==0.8.20.0
Pillow>=10.0.0
djangorestframework-simplejwt==5.3.1 